+++ src/mpvHandler.ts
@@ -16,7 +16,7 @@ export async function play(url: string) {
 
 export async function next() {
 	try {
-		await mpv.next("force");
+		await mpv.next();
 	}
 	catch (error) {
 		console.error("next ", error);
@@ -25,7 +25,7 @@ export async function next() {
 
 export async function prev() {
 	try {
-		await mpv.prev("force");
+		await mpv.prev();
 	}
 	catch (error) {
 		console.error("prev ", error);
@@ -97,6 +97,43 @@ export async function loadPlaylist(filePath: string) {
 		console.error("err loadPlaylist ==>", error);
 	}
 }
+
+export async function loopPlaylist(times: any) {
+	try {
+		await mpv.loopPlaylist(times);
+	}
+	catch (error) {
+		console.error("err loopPlaylist ==>", error);
+	}
+}
+
+export async function getPlaylistPosition() {
+	try {
+		await mpv.getPlaylistPosition();
+	}
+	catch (error) {
+		console.error("err getPlaylistPosition ==>", error);
+	}
+}
+
+export async function getTitle() {
+	try {
+		await mpv.getTitle();
+	}
+	catch (error) {
+		console.error("err getTitle ==>", error);
+	}
+}
+
+export async function observeProperty(property: String) {
+	try {
+		await mpv.observeProperty(property);
+	}
+	catch (error) {
+		console.error("err observeProperty property==>", error);
+	}
+}
+
 
 export async function quitMpvNeeded(): Promise<void> {
 	try {