+++ types/node-mpv.d.ts
@@ -104,6 +104,8 @@ declare class node_mpv {
     loadPlaylist(playlist: any, mode?: any): any;
 
     loop(times: any): any;
+    
+    loopPlaylist(times: any): any;
 
     messageHandler(message: any): void;
 
@@ -111,9 +113,9 @@ declare class node_mpv {
 
     mute(): any;
 
-    next(mode: any): any;
+    next(mode?: String): any;
 
-    observeProperty(property: any, id: any): void;
+    observeProperty(property: any): void;
 
     off(type: any, listener: any): any;
 
@@ -133,7 +135,7 @@ declare class node_mpv {
 
     prependOnceListener(type: any, listener: any): any;
 
-    prev(mode: any): any;
+    prev(mode?: String): any;
 
     quit(): void;